+++ database/java/migrations/ticket-order-db-migrations/ticket-order-analytical-migrations/src/main/resources/db/repeatable/R__event_details.sql
@@ -1,11 +1,10 @@
-CREATE OR REPLACE VIEW ticket_analytical.event_details AS
-SELECT
-    event_details_id,
-    event_id,
-    description,
-    number_of_places,
-    number_of_rows,
-    seats_per_row,
-    created_at,
-    updated_at
-FROM ticket_transactional.t_event_details;
+CREATE TABLE IF NOT EXISTS event_details (
+    event_details_id UUID PRIMARY KEY,
+    event_id UUID NOT NULL UNIQUE,
+    description TEXT NOT NULL,
+    number_of_places INTEGER NOT NULL,
+    number_of_rows INTEGER NOT NULL,
+    seats_per_row INTEGER NOT NULL,
+    created_at TIMESTAMP WITH TIME ZONE NOT NULL,
+    updated_at TIMESTAMP WITH TIME ZONE NOT NULL
+);