+++ database/java/migrations/ticket-order-db-migrations/ticket-order-transactional-migrations/src/main/resources/db/migrations/V1.0001__create_users.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS ticket_transactional.t_users (
+CREATE TABLE IF NOT EXISTS t_users (
     id UUID PRIMARY KEY,
     email VARCHAR(320) NOT NULL UNIQUE,
     password_hash VARCHAR(255) NOT NULL,