+++ database/java/migrations/ticket-order-db-migrations/ticket-order-transactional-migrations/src/main/resources/db/migrations/V1.0002__create_event.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS ticket_transactional.t_event (
+CREATE TABLE IF NOT EXISTS t_event (
     event_id UUID PRIMARY KEY,
     owner_id UUID NOT NULL,
     date TIMESTAMP WITH TIME ZONE NOT NULL,
@@ -10,10 +10,10 @@ CREATE TABLE IF NOT EXISTS ticket_transactional.t_event (
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL
 );
 
-ALTER TABLE ticket_transactional.t_event
+ALTER TABLE t_event
     ADD CONSTRAINT t_event_owner_id_fk
-        FOREIGN KEY (owner_id) REFERENCES ticket_transactional.t_users(id);
+        FOREIGN KEY (owner_id) REFERENCES t_users(id);
 
-ALTER TABLE ticket_transactional.t_event
+ALTER TABLE t_event
     ADD CONSTRAINT t_event_status_check
         CHECK (status IN ('DRAFT', 'PUBLISHED', 'CANCELLED'));