+++ database/java/migrations/ticket-order-db-migrations/ticket-order-transactional-migrations/src/main/resources/db/migrations/V1.0004__create_event_order.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS ticket_transactional.t_event_order (
+CREATE TABLE IF NOT EXISTS t_event_order (
     event_order_id UUID PRIMARY KEY,
     event_id UUID NOT NULL,
     customer_reference UUID NULL,
@@ -8,22 +8,22 @@ CREATE TABLE IF NOT EXISTS ticket_transactional.t_event_order (
     reservation_date TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
-ALTER TABLE ticket_transactional.t_event_order
+ALTER TABLE t_event_order
     ADD CONSTRAINT t_event_order_event_id_fk
-        FOREIGN KEY (event_id) REFERENCES ticket_transactional.t_event(event_id);
+        FOREIGN KEY (event_id) REFERENCES t_event(event_id);
 
-ALTER TABLE ticket_transactional.t_event_order
+ALTER TABLE t_event_order
     ADD CONSTRAINT t_event_order_row_number_check
         CHECK (row_number > 0);
 
-ALTER TABLE ticket_transactional.t_event_order
+ALTER TABLE t_event_order
     ADD CONSTRAINT t_event_order_place_number_check
         CHECK (place_number > 0);
 
-ALTER TABLE ticket_transactional.t_event_order
+ALTER TABLE t_event_order
     ADD CONSTRAINT t_event_order_place_type_check
         CHECK (length(btrim(place_type)) > 0);
 
-ALTER TABLE ticket_transactional.t_event_order
+ALTER TABLE t_event_order
     ADD CONSTRAINT t_event_order_event_position_unique
         UNIQUE (event_id, row_number, place_number);