+++ services/java/ticket-order-api/src/main/java/com/example/ticketplatform/api/adapter/out/persistence/user/UserEntity.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@Table(name = "t_users")
+@Table(name = "t_users", schema = "ticket_transactional")
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)